+++ NEWS.md
@@ -1,3 +1,10 @@
+# ReplicationSuccess 1.3.3
+
+- new functions `pEdgington`, `powerEdgington`, and `sampleSizeEdgington` to
+  design and analyze replication studies based on sum of *p*-values
+  (Edgington's) method (see <https://doi.org/10.48550/arXiv.2401.13615> for
+  details)
+
 # ReplicationSuccess 1.3.2
 
 - new maintainer Samuel Pawel